requirements.txt
linkify-it-py==2.1.0
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
platformdirs==4.9.6
plotext==5.3.2
Pygments==2.20.0
PyYAML==6.0.3
rich==14.3.4
textual==8.2.3
typing_extensions==4.15.0
uc-micro-py==2.0.0
